feat(join): allow a user to join a team, prevent posting to restricted or private teams

+++ backend/repos/base.py
@@ -72,5 +72,7 @@ class DatabaseAbstractRepository(AbstractRepository):
         self._db.commit()
 
     def delete(self, item: T) -> None:
+        if not item:
+            return
         self._db.delete(item)
         self._db.commit()
